Move constants and types that could shared in the package v1alpha1
