+++ pkg/controller/account/account_controller.go
@@ -2,7 +2,6 @@ package account
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"regexp"
 	"strconv"
@@ -62,89 +61,6 @@ const (
 
 	adminAccessArn = "arn:aws:iam::aws:policy/AdministratorAccess"
 )
-
-var coveredRegions = map[string]map[string]string{
-	"us-east-1": {
-		"initializationAMI": "ami-000db10762d0c4c05",
-	},
-	"us-east-2": {
-		"initializationAMI": "ami-094720ddca649952f",
-	},
-	"us-west-1": {
-		"initializationAMI": "ami-04642fc8fca1e8e67",
-	},
-	"us-west-2": {
-		"initializationAMI": "ami-0a7e1ebfee7a4570e",
-	},
-	"ca-central-1": {
-		"initializationAMI": "ami-06ca3c0058d0275b3",
-	},
-	"eu-central-1": {
-		"initializationAMI": "ami-09de4a4c670389e4b",
-	},
-	"eu-west-1": {
-		"initializationAMI": "ami-0202869bdd0fc8c75",
-	},
-	"eu-west-2": {
-		"initializationAMI": "ami-0188c0c5eddd2d032",
-	},
-	"eu-west-3": {
-		"initializationAMI": "ami-0c4224e392ec4e440",
-	},
-	"ap-northeast-1": {
-		"initializationAMI": "ami-00b95502a4d51a07e",
-	},
-	"ap-northeast-2": {
-		"initializationAMI": "ami-041b16ca28f036753",
-	},
-	"ap-south-1": {
-		"initializationAMI": "ami-0963937a03c01ecd4",
-	},
-	"ap-southeast-1": {
-		"initializationAMI": "ami-055c55112e25b1f1f",
-	},
-	"ap-southeast-2": {
-		"initializationAMI": "ami-036b423b657376f5b",
-	},
-	"sa-east-1": {
-		"initializationAMI": "ami-05c1c16cac05a7c0b",
-	},
-}
-
-// Custom errors
-
-// ErrAwsAccountLimitExceeded indicates the orgnization account limit has been reached.
-var ErrAwsAccountLimitExceeded = errors.New("AccountLimitExceeded")
-
-// ErrAwsInternalFailure indicates that there was an internal failure on the aws api
-var ErrAwsInternalFailure = errors.New("InternalFailure")
-
-// ErrAwsFailedCreateAccount indicates that an account creation failed
-var ErrAwsFailedCreateAccount = errors.New("FailedCreateAccount")
-
-// ErrAwsTooManyRequests indicates that to many requests were sent in a short period
-var ErrAwsTooManyRequests = errors.New("TooManyRequestsException")
-
-// ErrAwsCaseCreationLimitExceeded indicates that the support case limit for the account has been reached
-var ErrAwsCaseCreationLimitExceeded = errors.New("SupportCaseLimitExceeded")
-
-// ErrAwsFailedCreateSupportCase indicates that a support case creation failed
-var ErrAwsFailedCreateSupportCase = errors.New("FailedCreateSupportCase")
-
-// ErrAwsSupportCaseIDNotFound indicates that the support case ID was not found
-var ErrAwsSupportCaseIDNotFound = errors.New("SupportCaseIdNotfound")
-
-// ErrAwsFailedDescribeSupportCase indicates that the support case describe failed
-var ErrAwsFailedDescribeSupportCase = errors.New("FailedDescribeSupportCase")
-
-// ErrFederationTokenOutputNil indicates that getting a federation token from AWS failed
-var ErrFederationTokenOutputNil = errors.New("FederationTokenOutputNil")
-
-// ErrCreateEC2Instance indicates that the CreateEC2Instance function timed out
-var ErrCreateEC2Instance = errors.New("EC2CreationTimeout")
-
-// ErrFailedAWSTypecast indicates that there was a failure while typecasting to aws error
-var ErrFailedAWSTypecast = errors.New("FailedToTypecastAWSError")
 
 /**
 * USER ACTION REQUIRED: This is a scaffold file intended for the user to modify with their own Controller
@@ -412,8 +328,8 @@ func (r *ReconcileAccount) Reconcile(request reconcile.Request) (reconcile.Resul
 
 				// before doing anything make sure we are not over the limit if we are just error
 				if totalaccountwatcher.TotalAccountWatcher.Total >= AwsLimit {
-					reqLogger.Error(ErrAwsAccountLimitExceeded, "AWS Account limit reached", "Account Total", totalaccountwatcher.TotalAccountWatcher.Total)
-					return reconcile.Result{}, ErrAwsAccountLimitExceeded
+					reqLogger.Error(awsv1alpha1.ErrAwsAccountLimitExceeded, "AWS Account limit reached", "Account Total", totalaccountwatcher.TotalAccountWatcher.Total)
+					return reconcile.Result{}, awsv1alpha1.ErrAwsAccountLimitExceeded
 				}
 
 				// Build Aws Account
@@ -563,7 +479,7 @@ func (r *ReconcileAccount) Reconcile(request reconcile.Request) (reconcile.Resul
 		}
 
 		// Initialize all supported regions by creating and terminating an instance in each
-		err = r.InitializeSupportedRegions(reqLogger, coveredRegions, creds)
+		err = r.InitializeSupportedRegions(reqLogger, awsv1alpha1.CoveredRegions, creds)
 		if err != nil {
 			r.setStatusFailed(reqLogger, currentAcctInstance, "Failed to build and destroy ec2 instances")
 			return reconcile.Result{}, err
@@ -651,16 +567,16 @@ func (r *ReconcileAccount) BuildAccount(reqLogger logr.Logger, awsClient awsclie
 	// If it was an api or a limit issue don't modify account and exit if anything else set to failed
 	if orgErr != nil {
 		switch orgErr {
-		case ErrAwsFailedCreateAccount:
+		case awsv1alpha1.ErrAwsFailedCreateAccount:
 			SetAccountStatus(reqLogger, account, "Failed to create AWS Account", awsv1alpha1.AccountFailed, AccountFailed)
 			err := r.Client.Status().Update(context.TODO(), account)
 			if err != nil {
 				return "", err
 			}
 
-			reqLogger.Error(ErrAwsFailedCreateAccount, "Failed to create AWS Account")
+			reqLogger.Error(awsv1alpha1.ErrAwsFailedCreateAccount, "Failed to create AWS Account")
 			return "", orgErr
-		case ErrAwsAccountLimitExceeded:
+		case awsv1alpha1.ErrAwsAccountLimitExceeded:
 			log.Error(orgErr, "Failed to create AWS Account limit reached")
 			return "", orgErr
 		default:
@@ -709,13 +625,13 @@ func CreateAccount(reqLogger logr.Logger, client awsclient.Client, accountName, 
 		if aerr, ok := err.(awserr.Error); ok {
 			switch aerr.Code() {
 			case organizations.ErrCodeConstraintViolationException:
-				returnErr = ErrAwsAccountLimitExceeded
+				returnErr = awsv1alpha1.ErrAwsAccountLimitExceeded
 			case organizations.ErrCodeServiceException:
-				returnErr = ErrAwsInternalFailure
+				returnErr = awsv1alpha1.ErrAwsInternalFailure
 			case organizations.ErrCodeTooManyRequestsException:
-				returnErr = ErrAwsTooManyRequests
+				returnErr = awsv1alpha1.ErrAwsTooManyRequests
 			default:
-				returnErr = ErrAwsFailedCreateAccount
+				returnErr = awsv1alpha1.ErrAwsFailedCreateAccount
 				utils.LogAwsError(reqLogger, "New AWS Error during account creation", returnErr, err)
 			}
 
@@ -741,11 +657,11 @@ func CreateAccount(reqLogger logr.Logger, client awsclient.Client, accountName, 
 			var returnErr error
 			switch *status.CreateAccountStatus.FailureReason {
 			case "ACCOUNT_LIMIT_EXCEEDED":
-				returnErr = ErrAwsAccountLimitExceeded
+				returnErr = awsv1alpha1.ErrAwsAccountLimitExceeded
 			case "INTERNAL_FAILURE":
-				returnErr = ErrAwsInternalFailure
+				returnErr = awsv1alpha1.ErrAwsInternalFailure
 			default:
-				returnErr = ErrAwsFailedCreateAccount
+				returnErr = awsv1alpha1.ErrAwsFailedCreateAccount
 			}
 
 			return &organizations.DescribeCreateAccountStatusOutput{}, returnErr
